move isLaneSpeedClass to constants.ts
